+++ node/__tests__/__utils__/mocks/adServer.ts
@@ -9,6 +9,9 @@ export const getSponsoredProductsResponse: AdServerResponse = {
       campaignId: 'campaignId',
       adId: 'adId',
       actionCost: 0.32,
+      options: {
+        hideSponsoredBadge: true,
+      },
     },
   ],
 }